+++ src/DataPipes/CastPropertiesDataPipe.php
@@ -14,6 +14,7 @@ use Spatie\LaravelData\Support\Creation\CreationContext;
 use Spatie\LaravelData\Support\DataClass;
 use Spatie\LaravelData\Support\DataConfig;
 use Spatie\LaravelData\Support\DataProperty;
+use Spatie\LaravelData\Support\Types\CombinationType;
 
 class CastPropertiesDataPipe implements DataPipe
 {
@@ -97,8 +98,16 @@ class CastPropertiesDataPipe implements DataPipe
                 $creationContext->previous();
 
                 return $data;
-            } catch (CannotCreateData) {
-                return $value;
+            } catch (CannotCreateData $exception) {
+                $creationContext->previous();
+
+                if ($property->type->type instanceof CombinationType) {
+                    // Try another type in the union (which will need to be a simple type like string, int)
+                    // In the future we should deterministically choose the correct type to cast to
+                    return $value;
+                }
+
+                throw $exception;
             }
         }
 
@@ -137,10 +146,6 @@ class CastPropertiesDataPipe implements DataPipe
         array $properties,
         CreationContext $creationContext
     ): iterable {
-        if (empty($values)) {
-            return $values;
-        }
-
         if ($values instanceof Enumerable) {
             $values = $values->all();
         }
@@ -170,6 +175,10 @@ class CastPropertiesDataPipe implements DataPipe
         array $properties,
         CreationContext $creationContext
     ): array {
+        if(empty($values)) {
+            return $values;
+        }
+
         /** @var ?IterableItemCast $cast */
         $cast = $this->findCastForIterableItems($property, $values, $properties, $creationContext);
 
@@ -208,7 +217,7 @@ class CastPropertiesDataPipe implements DataPipe
             }
         }
 
-        if(in_array($property->type->iterableItemType, ['bool', 'int', 'float', 'array', 'string'])) {
+        if (in_array($property->type->iterableItemType, ['bool', 'int', 'float', 'array', 'string'])) {
             return new BuiltinTypeCast($property->type->iterableItemType);
         }